+++ lib/action_policy.rb
@@ -32,8 +32,9 @@ module ActionPolicy
     attr_accessor :cache_store
 
     # Find a policy class for a target
-    def lookup(target, allow_nil: false, **options)
+    def lookup(target, allow_nil: false, default: nil, **options)
       LookupChain.call(target, **options) ||
+        default ||
         (allow_nil ? nil : raise(NotFound, target))
     end
   end